+++ packages/core/src/runtime/world.ts
@@ -6,8 +6,13 @@ import { createVercelWorld } from '@workflow/world-vercel';
 
 const require = createRequire(join(process.cwd(), 'index.js'));
 
-let worldCache: World | undefined;
-let stubbedWorldCache: World | undefined;
+const WorldCache = Symbol.for('@workflow/world//cache');
+const StubbedWorldCache = Symbol.for('@workflow/world//stubbedCache');
+
+const globalSymbols: typeof globalThis & {
+  [WorldCache]?: World;
+  [StubbedWorldCache]?: World;
+} = globalThis;
 
 function defaultWorld(): 'vercel' | 'embedded' {
   if (process.env.VERCEL_DEPLOYMENT_ID) {
@@ -67,22 +72,22 @@ export const createWorld = (): World => {
  * be able to re-combine getWorld and getWorldHandlers into one singleton.
  */
 export const getWorldHandlers = (): Pick<World, 'createQueueHandler'> => {
-  if (stubbedWorldCache) {
-    return stubbedWorldCache;
+  if (globalSymbols[StubbedWorldCache]) {
+    return globalSymbols[StubbedWorldCache];
   }
   const _world = createWorld();
-  stubbedWorldCache = _world;
+  globalSymbols[StubbedWorldCache] = _world;
   return {
     createQueueHandler: _world.createQueueHandler,
   };
 };
 
 export const getWorld = (): World => {
-  if (worldCache) {
-    return worldCache;
+  if (globalSymbols[WorldCache]) {
+    return globalSymbols[WorldCache];
   }
-  worldCache = createWorld();
-  return worldCache;
+  globalSymbols[WorldCache] = createWorld();
+  return globalSymbols[WorldCache];
 };
 
 /**
@@ -90,6 +95,6 @@ export const getWorld = (): World => {
  * variables change and you need to reinitialize the world with new config.
  */
 export const setWorld = (world: World | undefined): void => {
-  worldCache = world;
-  stubbedWorldCache = world;
+  globalSymbols[WorldCache] = world;
+  globalSymbols[StubbedWorldCache] = world;
 };