+++ packages/world-postgres/src/drizzle/schema.ts
@@ -11,6 +11,7 @@ import {
   customType,
   index,
   integer,
+  /** @deprecated: use Cbor instead */
   jsonb,
   pgEnum,
   pgSchema,
@@ -19,6 +20,7 @@ import {
   timestamp,
   varchar,
 } from 'drizzle-orm/pg-core';
+import { Cbor, type Cborized } from './cbor.js';
 
 function mustBeMoreThanOne<T>(t: T[]) {
   return t as [T, ...T[]];
@@ -55,12 +57,19 @@ export const runs = schema.table(
   'workflow_runs',
   {
     runId: varchar('id').primaryKey(),
-    output: jsonb('output').$type<SerializedContent>(),
+    /** @deprecated */
+    outputJson: jsonb('output').$type<SerializedContent>(),
+    output: Cbor<SerializedContent>()('output_cbor'),
     deploymentId: varchar('deployment_id').notNull(),
     status: workflowRunStatus('status').notNull(),
     workflowName: varchar('name').notNull(),
-    executionContext: jsonb('execution_context').$type<Record<string, any>>(),
-    input: jsonb('input').$type<SerializedContent>().notNull(),
+    /** @deprecated */
+    executionContextJson:
+      jsonb('execution_context').$type<Record<string, any>>(),
+    executionContext: Cbor<Record<string, any>>()('execution_context_cbor'),
+    /** @deprecated */
+    inputJson: jsonb('input').$type<SerializedContent>(),
+    input: Cbor<SerializedContent>()('input_cbor'),
     error: text('error'),
     createdAt: timestamp('created_at').defaultNow().notNull(),
     updatedAt: timestamp('updated_at')
@@ -69,11 +78,13 @@ export const runs = schema.table(
       .notNull(),
     completedAt: timestamp('completed_at'),
     startedAt: timestamp('started_at'),
-  } satisfies DrizzlishOfType<WorkflowRun>,
-  (tb) => ({
-    workflowNameIdx: index().on(tb.workflowName),
-    statusIdx: index().on(tb.status),
-  })
+  } satisfies DrizzlishOfType<
+    Cborized<
+      Omit<WorkflowRun, 'input'> & { input?: unknown },
+      'input' | 'output' | 'executionContext'
+    >
+  >,
+  (tb) => [index().on(tb.workflowName), index().on(tb.status)]
 );
 
 export const events = schema.table(
@@ -84,12 +95,13 @@ export const events = schema.table(
     correlationId: varchar('correlation_id'),
     createdAt: timestamp('created_at').defaultNow().notNull(),
     runId: varchar('run_id').notNull(),
-    eventData: jsonb('payload'),
-  } satisfies DrizzlishOfType<Event & { eventData?: undefined }>,
-  (tb) => ({
-    runFk: index().on(tb.runId),
-    correlationIdFk: index().on(tb.correlationId),
-  })
+    /** @deprecated */
+    eventDataJson: jsonb('payload'),
+    eventData: Cbor<unknown>()('payload_cbor'),
+  } satisfies DrizzlishOfType<
+    Cborized<Event & { eventData?: undefined }, 'eventData'>
+  >,
+  (tb) => [index().on(tb.runId), index().on(tb.correlationId)]
 );
 
 export const steps = schema.table(
@@ -99,8 +111,12 @@ export const steps = schema.table(
     stepId: varchar('step_id').primaryKey(),
     stepName: varchar('step_name').notNull(),
     status: stepStatus('status').notNull(),
-    input: jsonb('input').$type<SerializedContent>().notNull(),
-    output: jsonb('output').$type<SerializedContent>(),
+    /** @deprecated */
+    inputJson: jsonb('input').$type<SerializedContent>(),
+    input: Cbor<SerializedContent>()('input_cbor'),
+    /** @deprecated we stream binary data */
+    outputJson: jsonb('output').$type<SerializedContent>(),
+    output: Cbor<SerializedContent>()('output_cbor'),
     error: text('error'),
     attempt: integer('attempt').notNull(),
     startedAt: timestamp('started_at'),
@@ -111,11 +127,10 @@ export const steps = schema.table(
       .$onUpdateFn(() => new Date())
       .notNull(),
     retryAfter: timestamp('retry_after'),
-  } satisfies DrizzlishOfType<Step>,
-  (tb) => ({
-    runFk: index().on(tb.runId),
-    statusIdx: index().on(tb.status),
-  })
+  } satisfies DrizzlishOfType<
+    Cborized<Omit<Step, 'input'> & { input?: unknown }, 'output' | 'input'>
+  >,
+  (tb) => [index().on(tb.runId), index().on(tb.status)]
 );
 
 export const hooks = schema.table(
@@ -128,12 +143,11 @@ export const hooks = schema.table(
     projectId: varchar('project_id').notNull(),
     environment: varchar('environment').notNull(),
     createdAt: timestamp('created_at').defaultNow().notNull(),
-    metadata: jsonb('metadata').$type<SerializedContent>(),
-  } satisfies DrizzlishOfType<Hook>,
-  (tb) => ({
-    runFk: index().on(tb.runId),
-    tokenIdx: index().on(tb.token),
-  })
+    /** @deprecated */
+    metadataJson: jsonb('metadata').$type<SerializedContent>(),
+    metadata: Cbor<SerializedContent>()('metadata_cbor'),
+  } satisfies DrizzlishOfType<Cborized<Hook, 'metadata'>>,
+  (tb) => [index().on(tb.runId), index().on(tb.token)]
 );
 
 const bytea = customType<{ data: Buffer; notNull: false; default: false }>({
@@ -151,7 +165,5 @@ export const streams = schema.table(
     createdAt: timestamp('created_at').defaultNow().notNull(),
     eof: boolean('eof').notNull(),
   },
-  (tb) => ({
-    primaryKey: primaryKey({ columns: [tb.streamId, tb.chunkId] }),
-  })
+  (tb) => [primaryKey({ columns: [tb.streamId, tb.chunkId] })]
 );