+++ workbench/vite/vite.config.ts
@@ -6,5 +6,6 @@ export default defineConfig({
   plugins: [nitro(), workflow()],
   nitro: {
     serverDir: './',
+    plugins: ['plugins/start-pg-world.ts'],
   },
 });