+++ packages/errors/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @workflow/errors
 
+## 4.0.1-beta.6
+
+### Patch Changes
+
+- Updated dependencies [edb69c3]
+  - @workflow/utils@4.0.1-beta.4
+
 ## 4.0.1-beta.5
 
 ### Patch Changes