+++ packages/nitro/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/workflow-nitro
 
+## 4.0.1-beta.21
+
+### Patch Changes
+
+- @workflow/core@4.0.1-beta.18
+- @workflow/builders@4.0.1-beta.17
+
+## 4.0.1-beta.20
+
+### Patch Changes
+
+- @workflow/core@4.0.1-beta.17
+- @workflow/builders@4.0.1-beta.16
+
 ## 4.0.1-beta.19
 
 ### Patch Changes